+++ src/interface.jl
@@ -417,6 +417,10 @@ function diff_left_compose(G::AbstractLieGroup, g, h, X)
     Y = ManifoldsBase.allocate_result(G, diff_left_compose, X, g, h)
     return diff_left_compose!(G, Y, g, h, X)
 end
+function diff_left_compose(G::AbstractLieGroup{𝔽, O}, g::Identity{O}, h, X) where {𝔽, O <: AbstractGroupOperation}
+    Y = ManifoldsBase.allocate_result(G, diff_left_compose, X, h)
+    return diff_left_compose!(G, Y, g, h, X)
+end
 
 function diff_left_compose! end
 @doc "$(_doc_diff_left_compose)"
@@ -453,6 +457,11 @@ $(_math(:d)) λ_g(h)[X] = X ∈ 𝔤.
 @doc "$(_doc_diff_right_compose)"
 function diff_right_compose(G::AbstractLieGroup, g, h, X)
     Y = ManifoldsBase.allocate_result(G, diff_right_compose, X, g, h)
+    return diff_right_compose!(G, Y, g, h, X)
+end
+
+function diff_right_compose(G::AbstractLieGroup{𝔽, O}, g::Identity{O}, h, X) where {𝔽, O <: AbstractGroupOperation}
+    Y = ManifoldsBase.allocate_result(G, diff_right_compose, X, h)
     return diff_right_compose!(G, Y, g, h, X)
 end